+++ resources/assets/js/components/products/orders/order.js
@@ -7,9 +7,13 @@ export default {
   },
   methods: {
     deleteOrder () {
-      axios.delete(`/webapi/orders/${this.order.id}`).then((response) => {
-        this.$emit('orderDeleted')
-      })
+      if (confirm(`Вы уверены, что хотите удалить заказ № ${this.order.id}?`)) {
+        axios.delete(`/webapi/orders/${this.order.id}`).then((response) => {
+          this.$emit('orderDeleted')
+        })
+      } else {
+        // Do nothing!
+      }
     },
   },
   computed: {
